Commit 006: input and generalization
Commit 006: input and generalization

Refactored the main menu to take up to 20 side lengths or coordinates,
and determine the type of polygon from there.  The getTriangleSides()
function was refactored into inputSideLengths() for this purpose.  Also,
the switch statement in main was moved to a separate interface function.
I also added a pointsMatch() function for checking (x,y) coordinates
input.